Removing old code

+++ src/rand.h
@@ -5,10 +5,10 @@ using namespace std;
 
 namespace DSMCpp {
     namespace Random {
-        double rand();
-        double rand(double v0, double v1);
-        double randn();
-        double randn(double mu, double sigma);
+        float rand();
+        float rand(float v0, float v1);
+        float randn();
+        float randn(float mu, float sigma);
         int rand_int(float v0, float v1);
     }
 }
